Finance/01-dynamic-pricing/azure-deployment/functions/requirements.txt
azure-functions
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
azure-storage-blob>=12.14.0
pyodbc>=4.0.0
